Restructured the project to categorise the implementation and test code under the respective challenge_id
