+++ src/BeFaster.App.Tests/BeFaster.App.Tests.csproj
@@ -40,7 +40,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Solutions\SumSolutionTest.cs" />
+    <Compile Include="Solutions\SUM\SumSolutionTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
