+++ src/BeFaster.App.Tests/Solutions/SUM/SumSolutionTest.cs
@@ -1,7 +1,7 @@
 ﻿using BeFaster.App.Solutions;
 using NUnit.Framework;
 
-namespace BeFaster.App.Tests.Solutions
+namespace BeFaster.App.Tests.Solutions.SUM
 {
     [TestFixture]
     public class SumSolutionTest
