+++ src/BeFaster.App/Solutions/SUM/SumSolution.cs
@@ -1,6 +1,6 @@
 ﻿using BeFaster.Runner.Exceptions;
 
-namespace BeFaster.App.Solutions
+namespace BeFaster.App.Solutions.SUM
 {
     public static class SumSolution
     {
